[IMP] props: interface for selecting props dynamically
This commit introduces a props panel that allows users to interact with
props dynamically. The panel allows users to modify the values of props
on the fly, without the need to restart the application.
This provides a more seamless and efficient experience for users when
working with props in the application.

The supported props for dynamic modification are for the moment `String`
and `Boolean`. Other type of props will be implemented in the future.

This commit also change the way stories are represented.
Before a story was a `Component` class that rendered the component, with
static objects to represent the story metadata.
Now a story is a simple object where each attribute represent the story
metadata.