[IMP] adding search bar
This PR add a search bar to the sidebar. When typing in the searchbar, the element that doesn't match the entry are not displayed (this works for stories and folder but not yet on the highest level of stories).
Also this PR improve the xml/css of the sidebar so that the searchbar fit nicely.