[IMP] Move ui_playground to a Controller (#13)
By moving the ui_playground into a controller, external users will
be able to access the app without logging in. Additionally, this
may also enable the integration of the ui_playground's UI via an
iframe on other websites.

This PR is far from being perfect.
Assets in Odoo is something I find rather complex. 
Especially when you want something very specific like having a
owl app in a controller while having access to the backend assets.

For now, it works. But I feel that we load more assets than 
necessary. We will come back to it if the time allow us to
do so