[IMP] Introduce views support
This commit adds a new type of stories, stories that renders arch views.
Thanks to this kind of stories, users can now play with views and their
archs.

There is also a code editor available to the users to edit the arch
directly in the browser and directly see the result.

This commits also simplify the assets.