[IMP] add custom URL for stories
The purpose of this PR is to add a custom URL when being in a story. The string "#module=moduleName&folder=folderName&story=storyName" is added to the base URL. In addition, we needed to handle the rollback so that when the back button is pressed, the right stories become active. Finally, the last thing to do is to take care of the refresh, when we do a refresh, the component is rendered again and we would go back to the introduction, but since the URL was not changed, we can easily change the active story so that the user goes back to the page he was on before the refresh.