[REF & FIX] narrow down the bundle & remove QUnit
Previously, we didn't removed QUnit from our assets because some
files we are using depends on it.
With this commit, QUnit is deleted from our assets and the API is
mocked.

I also took the opportunity to narrow down our assets to not include
useless files we are not using.

As a side effect from this REF, some errors throwed by QUnit are
now fixed.